added loginpage and two folders for assets and testing, also changed the pubspec file to include the new assets
